feat: refine network selector
